dev: Restructure `cmd` package utils
